# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * The Child Lost Agency API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Case {
  id: number;
  caseId?: string;
  reporterName?: string;
  reporterPhone?: string;
  childName: string;
  childAge: number;
  country?: string;
  province?: string;
  lastSeen?: string;
  dateMissing?: string;
  description?: string;
  gpsEnrolled?: string;
  caseType: string;
  status: string;
  submittedAt: string;
}

export interface CaseInput {
  reporterName: string;
  reporterPhone: string;
  childName: string;
  childAge: number;
  country?: string;
  province?: string;
  lastSeen: string;
  dateMissing: string;
  description: string;
  gpsEnrolled?: string;
  caseType: string;
}

export interface CaseUpdate {
  status?: string;
}

export type CaseStatsByTypeItem = {
  type: string;
  count: number;
};

export interface CaseStats {
  total: number;
  open: number;
  resolved: number;
  gpsEnrolled?: number;
  byType: CaseStatsByTypeItem[];
}

export type AiChatInputMessagesItem = {
  role: string;
  content: string;
};

export interface AiChatInput {
  messages: AiChatInputMessagesItem[];
}

export interface ChatReply {
  reply: string;
}

export type AiCaseAnalysisInputCaseData = {[key: string]: string};

export interface AiCaseAnalysisInput {
  caseData: AiCaseAnalysisInputCaseData;
}

export interface AiAlertInput {
  situation?: string;
  area?: string;
}

export interface AiMaintenanceInput {
  systems?: string[];
  notes?: string;
}

export type AiBatteryInputUnitsItem = {
  name?: string;
  charge?: number;
  type?: string;
};

export interface AiBatteryInput {
  units?: AiBatteryInputUnitsItem[];
}

export interface AiThreatInput {
  target?: string;
  weapons?: string[];
  location?: string;
}

export interface AiTextResponse {
  text: string;
}