# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Child Lost Agency API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cases
    description: Missing child case operations
  - name: ai
    description: AI-powered analysis endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Submit a new missing child case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInput"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /cases/stats:
    get:
      operationId: getCaseStats
      tags: [cases]
      summary: Get case statistics summary
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStats"

  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a specific case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Not found
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update case status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: CLA Virtual Agent chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatReply"

  /ai/case-analysis:
    post:
      operationId: aiCaseAnalysis
      tags: [ai]
      summary: Analyze a missing child case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCaseAnalysisInput"
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTextResponse"

  /ai/emergency-alert:
    post:
      operationId: aiEmergencyAlert
      tags: [ai]
      summary: Generate emergency alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiAlertInput"
      responses:
        "200":
          description: Alert text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTextResponse"

  /ai/maintenance:
    post:
      operationId: aiMaintenance
      tags: [ai]
      summary: Generate system maintenance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMaintenanceInput"
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTextResponse"

  /ai/battery:
    post:
      operationId: aiBattery
      tags: [ai]
      summary: Analyze fleet battery and power levels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiBatteryInput"
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTextResponse"

  /ai/threat:
    post:
      operationId: aiThreat
      tags: [ai]
      summary: Generate threat assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiThreatInput"
      responses:
        "200":
          description: Assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiTextResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Case:
      type: object
      required: [id, childName, childAge, caseType, status, submittedAt]
      properties:
        id:
          type: integer
        caseId:
          type: string
        reporterName:
          type: string
        reporterPhone:
          type: string
        childName:
          type: string
        childAge:
          type: integer
        country:
          type: string
        province:
          type: string
        lastSeen:
          type: string
        dateMissing:
          type: string
        description:
          type: string
        gpsEnrolled:
          type: string
        caseType:
          type: string
        status:
          type: string
        submittedAt:
          type: string

    CaseInput:
      type: object
      required: [childName, childAge, caseType, lastSeen, dateMissing, description, reporterName, reporterPhone]
      properties:
        reporterName:
          type: string
        reporterPhone:
          type: string
        childName:
          type: string
        childAge:
          type: integer
        country:
          type: string
        province:
          type: string
        lastSeen:
          type: string
        dateMissing:
          type: string
        description:
          type: string
        gpsEnrolled:
          type: string
        caseType:
          type: string

    CaseUpdate:
      type: object
      properties:
        status:
          type: string

    CaseStats:
      type: object
      required: [total, open, resolved, byType]
      properties:
        total:
          type: integer
        open:
          type: integer
        resolved:
          type: integer
        gpsEnrolled:
          type: integer
        byType:
          type: array
          items:
            type: object
            required: [type, count]
            properties:
              type:
                type: string
              count:
                type: integer

    AiChatInput:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
              content:
                type: string

    ChatReply:
      type: object
      required: [reply]
      properties:
        reply:
          type: string

    AiCaseAnalysisInput:
      type: object
      required: [caseData]
      properties:
        caseData:
          type: object
          additionalProperties:
            type: string

    AiAlertInput:
      type: object
      properties:
        situation:
          type: string
        area:
          type: string

    AiMaintenanceInput:
      type: object
      properties:
        systems:
          type: array
          items:
            type: string
        notes:
          type: string

    AiBatteryInput:
      type: object
      properties:
        units:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              charge:
                type: integer
              type:
                type: string

    AiThreatInput:
      type: object
      properties:
        target:
          type: string
        weapons:
          type: array
          items:
            type: string
        location:
          type: string

    AiTextResponse:
      type: object
      required: [text]
      properties:
        text:
          type: string