# lib/api-zod/src/generated/types/aiAlertInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * The Child Lost Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface AiAlertInput {
  situation?: string;
  area?: string;
}