# lib/api-zod/src/generated/types/aiThreatInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * The Child Lost Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface AiThreatInput {
  target?: string;
  weapons?: string[];
  location?: string;
}