# lib/api-zod/src/generated/types/caseInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * The Child Lost Agency API
 * OpenAPI spec version: 0.1.0
 */

export interface CaseInput {
  reporterName: string;
  reporterPhone: string;
  childName: string;
  childAge: number;
  country?: string;
  province?: string;
  lastSeen: string;
  dateMissing: string;
  description: string;
  gpsEnrolled?: string;
  caseType: string;
}