# lib/api-zod/src/generated/types/caseStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * The Child Lost Agency API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatsByTypeItem } from './caseStatsByTypeItem';

export interface CaseStats {
  total: number;
  open: number;
  resolved: number;
  gpsEnrolled?: number;
  byType: CaseStatsByTypeItem[];
}